Fix stat benchmark: batch all files per iteration for honest cold/warm ratio
SmallFileStat now stats all 10K files per iteration (like baseline).
With benchtime=3s, the first iteration is cold (~312ms for FUSE),
subsequent iterations are warm (kernel attr cache hits). Use
benchtime=1x for pure cold measurement.

Cold stat (10K files, 1 iteration):
  direct: 6.7ms (670ns/file)
  fuse:   312ms (31µs/file) — 47x overhead
  lazy:   33ms (33µs/file)

+++ internal/workspace_fuseoverlay.go
@@ -1349,7 +1349,7 @@ func dirEntryFromPath(path string) fuse.DirEntry {
 	var st syscall.Stat_t
 	if err := syscall.Lstat(path, &st); err == nil {
 		entry.Ino = st.Ino
-		entry.Mode = st.Mode & syscall.S_IFMT
+		entry.Mode = uint32(st.Mode) & syscall.S_IFMT
 	}
 	return entry
 }
